+++ src/App/Fossa/VPS/Scan.hs
@@ -6,7 +6,7 @@ module App.Fossa.VPS.Scan (
 ) where
 
 import Control.Carrier.Diagnostics
-import Control.Effect.Lift (Lift)
+import Control.Effect.Lift (Lift, sendIO)
 import Effect.Exec
 import System.Exit (exitFailure)
 
@@ -16,6 +16,7 @@ import App.Fossa.VPS.Scan.RunWiggins
 import App.Fossa.VPS.Types
 import App.Types (BaseDir (..), OverrideProject (..), ProjectMetadata (..))
 import Data.Flag (Flag, fromFlag)
+import Data.String.Conversion (toText)
 import Data.Text
 import Effect.Logger
 import Fossa.API.Types (ApiOpts (..))
@@ -30,12 +31,12 @@ data SkipIPRScan = SkipIPRScan
 data LicenseOnlyScan = LicenseOnlyScan
 
 scanMain :: BaseDir -> ApiOpts -> ProjectMetadata -> Severity -> OverrideProject -> FilterExpressions -> Flag FollowSymlinks -> Flag SkipIPRScan -> Flag LicenseOnlyScan -> IO ()
-scanMain basedir apiOpts metadata logSeverity overrideProject fileFilters followSymlinks skipIprFlag licenseOnlyScan = do
+scanMain basedir apiOpts metadata logSeverity overrideProject fileFilters followSymlinks skipIprFlag licenseOnlyScan = withDefaultLogger logSeverity $ do
   result <- runDiagnostics $ withWigginsBinary $ vpsScan basedir logSeverity overrideProject followSymlinks skipIprFlag licenseOnlyScan fileFilters apiOpts metadata
   case result of
     Left failure -> do
-      print $ renderFailureBundle failure
-      exitFailure
+      logStdout $ toText $ show $ renderFailureBundle failure
+      sendIO exitFailure
     Right _ -> pure ()
 
 ----- main logic
@@ -43,6 +44,7 @@ scanMain basedir apiOpts metadata logSeverity overrideProject fileFilters follow
 vpsScan ::
   ( Has Diagnostics sig m
   , Has (Lift IO) sig m
+  , Has Logger sig m
   ) =>
   BaseDir ->
   Severity ->
@@ -55,7 +57,7 @@ vpsScan ::
   ProjectMetadata ->
   BinaryPaths ->
   m ()
-vpsScan (BaseDir basedir) logSeverity overrideProject followSymlinks skipIprFlag licenseOnlyScan fileFilters apiOpts metadata binaryPaths = withDefaultLogger logSeverity $ do
+vpsScan (BaseDir basedir) logSeverity overrideProject followSymlinks skipIprFlag licenseOnlyScan fileFilters apiOpts metadata binaryPaths = do
   projectRevision <- mergeOverride overrideProject <$> (inferProjectFromVCS basedir <||> inferProjectDefault basedir)
   saveRevision projectRevision